requirements.txt
fastapi
uvicorn
gunicorn
gspread
requests
python-dotenv
python-multipart
google-auth